feat(consumer): abstract + example:q
